+++ src/Elders.Cronus.Projections.Cassandra/EventSourcing/IHaveState.cs
@@ -1,8 +1,0 @@
-﻿namespace Elders.Cronus.Projections.Cassandra.EventSourcing
-{
-    public interface IHaveState
-    {
-        object State { get; set; }
-        void InitializeState(object state);
-    }
-}