+++ src/Elders.Cronus.Projections.Cassandra/EventSourcing/IProjectionStore.cs
@@ -1,45 +1,15 @@
-﻿using Elders.Cronus.DomainModeling;
-using System;
+﻿using System;
+using Elders.Cronus.DomainModeling;
+using Elders.Cronus.Projections.Cassandra.Snapshots;
 
 namespace Elders.Cronus.Projections.Cassandra.EventSourcing
 {
-    public interface IProjectionRegistryStore
-    {
-        void Save();
-        void Get();
-    }
-
-    public class ProjectionRegistryStore : IProjectionRegistryStore
-    {
-        public void Get()
-        {
-            throw new NotImplementedException();
-        }
-
-        public void Save()
-        {
-            throw new NotImplementedException();
-        }
-    }
-
-    public interface IProjectionRegistry
-    {
-
-    }
-
-    public class DataContractProjectionRegistry : IProjectionRegistry
-    {
-        public void RegisterProjection(IProjection projection)
-        {
-
-        }
-    }
-
     public interface IProjectionStore
     {
-        ProjectionStream Load(Type projectionType, IBlobId projectionId, ISnapshot snapshot);
-        ProjectionStream Load<T>(IBlobId projectionId, ISnapshot snapshot) where T : IProjectionDefinition;
+        ProjectionStream Load(string contractId, IBlobId projectionId, ISnapshot snapshot);
 
         void Save(ProjectionCommit commit);
+
+        IProjectionBuilder GetBuilder(Type projectionType);
     }
 }