+++ src/Elders.Cronus.Projections.Cassandra/EventSourcing/ProjectionStream.cs
@@ -1,5 +1,8 @@
-﻿using System.Collections.Generic;
+﻿using Elders.Cronus.Projections.Cassandra.Snapshots;
+using System.Collections.Generic;
 using System.Linq;
+using System;
+using Elders.Cronus.DomainModeling.Projections;
 
 namespace Elders.Cronus.Projections.Cassandra.EventSourcing
 {
@@ -10,6 +13,9 @@ namespace Elders.Cronus.Projections.Cassandra.EventSourcing
 
         public ProjectionStream(IList<ProjectionCommit> commits, ISnapshot snapshot)
         {
+            if (ReferenceEquals(null, commits) == true) throw new ArgumentException(nameof(commits));
+            if (ReferenceEquals(null, snapshot) == true) throw new ArgumentException(nameof(snapshot));
+
             this.snapshot = snapshot;
             this.commits = commits;
         }
@@ -22,6 +28,10 @@ namespace Elders.Cronus.Projections.Cassandra.EventSourcing
             if (events.Count > 0)
             {
                 var projection = (T)FastActivator.CreateInstance(typeof(T), true);
+
+                if (ReferenceEquals(null, snapshot.State) == false)
+                    projection.State = snapshot.State;
+
                 projection.ReplayEvents(events);
                 return new ProjectionGetResult<T>(true, projection);
             }