+++ src/Elders.Cronus.Projections.Cassandra/EventSourcing/Usage.cs
@@ -1,5 +1,6 @@
 ﻿using Elders.Cronus.DomainModeling;
 using System.Runtime.Serialization;
+using Elders.Cronus.DomainModeling.Projections;
 
 namespace Elders.Cronus.Projections.Cassandra.EventSourcing
 {
@@ -57,8 +58,6 @@ namespace Elders.Cronus.Projections.Cassandra.EventSourcing
         {
             IProjectionRepository repo = null;
             var proj = repo.Get<UserProjection>(null).Projection.State;
-
-
         }
     }
 }