+++ src/Elders.Cronus.Projections.Cassandra/Snapshots/ISnapshotStore.cs
@@ -1,0 +1,11 @@
+﻿using Elders.Cronus.DomainModeling;
+
+namespace Elders.Cronus.Projections.Cassandra.Snapshots
+{
+    public interface ISnapshotStore
+    {
+        ISnapshot Load(string projectionContractId, IBlobId id);
+
+        void Save(ISnapshot snapshot);
+    }
+}